+++ tweakwcs/tests/test_matchutils.py
@@ -17,7 +17,8 @@ from astropy.utils.data import get_pkg_data_filename
 
 import tweakwcs
 from tweakwcs.matchutils import (_xy_2dhist, _estimate_2dhist_shift,
-                                 _find_peak, XYXYMatch, MatchCatalogs)
+                                 _find_peak, XYXYMatch, MatchCatalogs,
+                                 MatchSourceConfusionError)
 from .helper_correctors import DummyWCSCorrector
 
 
@@ -291,6 +292,14 @@ def test_tpmatch(tp_wcs, use2dhist):
         tp_pscale=tp_pscale,
         tp_units=None if tp_wcs is None else tp_wcs.units
     )
+
+
+def test_multi_match_error():
+    tpmatch = XYXYMatch(tolerance=1.0, separation=0.01)
+    refcat = Table([[0.0, 0.1], [0.1, 0.0]], names=('TPx', 'TPy'), meta={'name': None})
+    imcat = Table([[0.0], [0.0]], names=('TPx', 'TPy'), meta={'name': None})
+    with pytest.raises(MatchSourceConfusionError):
+        tpmatch(refcat, imcat)
 
 
 def test_match_catalogs_abc():